refactor: move time-related types to separate module
