+++ src/specification.rs
@@ -2,6 +2,8 @@
 //! [specification](https://popsim-consortium.github.io/demes-spec-docs/main/specification.html)
 //! in terms of rust structs.
 
+use crate::time::*;
+use crate::traits::Validate;
 use crate::DemesError;
 use serde::{Deserialize, Serialize};
 use std::collections::{HashMap, HashSet};
@@ -17,201 +19,6 @@ macro_rules! get_deme {
         }
     };
 }
-
-/// Store time values.
-///
-/// This is a newtype wrapper for [`f64`](std::primitive::f64).
-///
-/// # Notes
-///
-/// * The units are in the [`TimeUnits`](crate::TimeUnits)
-///   of the [`Graph`](crate::Graph).
-/// * Invalid values are caught when a `Graph` is
-///   resolved.  Funcions that generate resolved graphs are:
-///    - [`loads`](crate::loads)
-///    - [`load`](crate::load)
-///    - [`GraphBuilder::resolve`](crate::GraphBuilder::resolve)
-///
-/// # Examples
-///
-/// ## In a `YAML` record
-///
-/// ```
-/// let yaml = "
-/// time_units: years
-/// generation_time: 25
-/// description: A deme that existed until 20 years ago.
-/// demes:
-///  - name: deme
-///    epochs:
-///     - start_size: 50
-///       end_time: 20
-/// ";
-/// demes::loads(yaml).unwrap();
-/// ```
-///
-/// ## Using rust code
-///
-/// Normally, one only needs to create a `Time` when
-/// working with [`GraphBuilder`](crate::GraphBuilder).
-///
-/// ```
-/// let t = demes::Time::from(0.0);
-/// assert_eq!(t, 0.0);
-/// ```
-#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
-#[repr(transparent)]
-#[serde(try_from = "TimeTrampoline")]
-#[serde(into = "TimeTrampoline")]
-pub struct Time(f64);
-
-impl Time {
-    fn default_deme_start_time() -> Self {
-        Self(f64::INFINITY)
-    }
-
-    fn default_epoch_end_time() -> Self {
-        Self(0.0)
-    }
-
-    fn is_valid_deme_start_time(&self) -> bool {
-        self.0 > 0.0
-    }
-
-    pub(crate) fn err_if_not_valid_deme_start_time(&self) -> Result<(), DemesError> {
-        if self.is_valid_deme_start_time() {
-            Ok(())
-        } else {
-            let msg = format!("start_time must be > 0.0, got: {}", self.0);
-            Err(DemesError::DemeError(msg))
-        }
-    }
-
-    fn is_valid_epoch_end_time(&self) -> bool {
-        self.0.is_finite()
-    }
-
-    fn err_if_not_valid_epoch_end_time(&self) -> Result<(), DemesError> {
-        if self.is_valid_epoch_end_time() {
-            Ok(())
-        } else {
-            let msg = format!("end_time must be <= t < Infinity, got: {}", self.0);
-            Err(DemesError::EpochError(msg))
-        }
-    }
-
-    fn is_valid_pulse_time(&self) -> bool {
-        self.0.is_sign_positive() && !self.0.is_infinite()
-    }
-
-    fn validate<F>(&self, f: F) -> Result<(), DemesError>
-    where
-        F: std::ops::FnOnce(String) -> DemesError,
-    {
-        if self.0.is_nan() || self.0.is_sign_negative() {
-            Err(f(format!("invalid time value: {}", self.0)))
-        } else {
-            Ok(())
-        }
-    }
-}
-
-impl_newtype_traits!(Time);
-
-/// Specify rounding method for
-/// [`Graph::to_integer_generations`](crate::Graph::to_integer_generations)
-#[derive(Copy, Clone)]
-pub enum RoundTimeToInteger {
-    /// Use [`f64::round`](std::primitive::f64::round)
-    F64,
-}
-
-impl RoundTimeToInteger {
-    fn apply_rounding(&self, time: Time, generation_time: GenerationTime) -> Time {
-        let mut temp_time = time.0 / generation_time.0;
-
-        match self {
-            RoundTimeToInteger::F64 => {
-                temp_time = temp_time.round();
-            }
-        }
-        Time(temp_time)
-    }
-}
-
-// Workhorse behing Graph::to_generations
-fn convert_resolved_time_to_generations<F>(
-    generation_time: GenerationTime,
-    rounding: Option<RoundTimeToInteger>,
-    f: F,
-    message: &str,
-    input: Option<Time>,
-) -> Result<Time, DemesError>
-where
-    F: std::ops::FnOnce(String) -> DemesError,
-{
-    match input {
-        Some(value) => match rounding {
-            Some(rounding_policy) => Ok(rounding_policy.apply_rounding(value, generation_time)),
-            None => Ok(Time::from(value.0 / generation_time.0)),
-        },
-        None => Err(f(message.to_string())),
-    }
-}
-
-#[derive(Serialize, Deserialize)]
-#[serde(untagged)]
-enum TimeTrampoline {
-    Infinity(String),
-    Float(f64),
-}
-
-impl TryFrom<TimeTrampoline> for Time {
-    type Error = DemesError;
-
-    fn try_from(value: TimeTrampoline) -> Result<Self, Self::Error> {
-        match value {
-            // Handle string inputs
-            TimeTrampoline::Infinity(string) => {
-                if &string == "Infinity" {
-                    Ok(Self(f64::INFINITY))
-                } else {
-                    Err(DemesError::GraphError(string))
-                }
-            }
-            // Fall back to valid YAML representations
-            TimeTrampoline::Float(f) => Ok(Self::from(f)),
-        }
-    }
-}
-
-impl From<Time> for TimeTrampoline {
-    fn from(value: Time) -> Self {
-        if value.0.is_infinite() {
-            Self::Infinity("Infinity".to_string())
-        } else {
-            Self::Float(f64::from(value))
-        }
-    }
-}
-
-#[repr(transparent)]
-struct HashableTime(Time);
-
-impl std::hash::Hash for HashableTime {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        let value = f64::from(self.0);
-        value.to_bits().hash(state)
-    }
-}
-
-impl PartialEq for HashableTime {
-    fn eq(&self, other: &Self) -> bool {
-        self.0.eq(&other.0)
-    }
-}
-
-impl Eq for HashableTime {}
 
 /// The size of a [`Deme`](crate::Deme) at a given [`Time`](crate::Time).
 ///
@@ -384,7 +191,7 @@ impl TimeInterval {
         F: Into<f64>,
     {
         let time = other.into();
-        self.start_time.0 > time && time >= self.end_time.0
+        self.start_time > time && time >= self.end_time
     }
 
     // true if other is in (start_time, end_time]
@@ -394,7 +201,7 @@ impl TimeInterval {
     {
         let time = other.into();
 
-        time > self.end_time.0 && time <= self.start_time.0
+        time > self.end_time && time <= self.start_time
     }
 
     fn contains_exclusive_start_inclusive_end<F>(&self, other: F) -> bool
@@ -403,7 +210,7 @@ impl TimeInterval {
     {
         let time = other.into();
 
-        time >= self.end_time.0 && time < self.start_time.0
+        time >= self.end_time && time < self.start_time
     }
 
     fn contains_inclusive<F>(&self, other: F) -> bool
@@ -411,7 +218,7 @@ impl TimeInterval {
         F: Into<f64>,
     {
         let time = other.into();
-        self.start_time.0 >= time && time >= self.end_time.0
+        self.start_time >= time && time >= self.end_time
     }
 
     fn duration_greater_than_zero(&self) -> bool {
@@ -978,7 +785,7 @@ impl UnresolvedPulse {
         if !time.is_valid_pulse_time() {
             return Err(DemesError::PulseError(format!(
                 "invalid pulse time: {}",
-                time.0
+                f64::from(time)
             )));
         }
 
@@ -2154,76 +1961,6 @@ fn deme_name_exists<F: FnOnce(String) -> DemesError>(
     }
 }
 
-/// The time units of a graph
-#[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
-#[serde(from = "String")]
-#[serde(into = "String")]
-pub enum TimeUnits {
-    #[allow(missing_docs)]
-    Generations,
-    #[allow(missing_docs)]
-    Years,
-    /// A "custom" time unit.  It is assumed
-    /// that client code knows what to do with this.
-    Custom(String),
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Serialize)]
-#[repr(transparent)]
-struct CustomTimeUnits(String);
-
-impl From<String> for TimeUnits {
-    fn from(value: String) -> Self {
-        if &value == "generations" {
-            Self::Generations
-        } else if &value == "years" {
-            Self::Years
-        } else {
-            Self::Custom(value)
-        }
-    }
-}
-
-impl From<TimeUnits> for String {
-    fn from(value: TimeUnits) -> Self {
-        value.to_string()
-    }
-}
-
-impl std::fmt::Display for TimeUnits {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            TimeUnits::Generations => write!(f, "generations"),
-            TimeUnits::Years => write!(f, "years"),
-            TimeUnits::Custom(custom) => write!(f, "{}", &custom),
-        }
-    }
-}
-
-/// Generation time.
-///
-/// If [`TimeUnits`] are in generations, this value
-/// must be 1.0.
-#[derive(Clone, Copy, Debug, Serialize, Deserialize)]
-#[repr(transparent)]
-#[serde(from = "f64")]
-pub struct GenerationTime(f64);
-
-impl GenerationTime {
-    fn validate(&self) -> Result<(), DemesError> {
-        if !self.0.is_finite() || !self.0.is_sign_positive() || !self.gt(&0.0) {
-            Err(DemesError::GraphError(format!(
-                "generation time must be > 0.0, got: {}",
-                self
-            )))
-        } else {
-            Ok(())
-        }
-    }
-}
-
-impl_newtype_traits!(GenerationTime);
-
 #[derive(Default, Debug, Serialize, Deserialize)]
 #[serde(deny_unknown_fields)]
 struct GraphDefaultInput {
@@ -2749,17 +2486,17 @@ impl UnresolvedGraph {
     fn get_non_overlapping_migration_intervals(&self) -> Vec<TimeInterval> {
         let mut unique_times = HashSet::<HashableTime>::default();
         for migration in &self.resolved_migrations {
-            unique_times.insert(HashableTime(migration.start_time()));
-            unique_times.insert(HashableTime(migration.end_time()));
+            unique_times.insert(HashableTime::from(migration.start_time()));
+            unique_times.insert(HashableTime::from(migration.end_time()));
         }
-        unique_times.retain(|t| f64::from(t.0).is_finite());
+        unique_times.retain(|t| f64::from(*t).is_finite());
 
-        let mut end_times = unique_times.into_iter().map(|x| x.0).collect::<Vec<_>>();
+        let mut end_times = unique_times.into_iter().map(Time::from).collect::<Vec<_>>();
 
         // REVERSE sort
         end_times.sort_by(|a, b| b.cmp(a));
 
-        let mut start_times = vec![Time(f64::INFINITY)];
+        let mut start_times = vec![Time::from(f64::INFINITY)];
 
         if let Some((_last, elements)) = end_times.split_last() {
             start_times.extend_from_slice(elements);
@@ -2844,7 +2581,7 @@ impl UnresolvedGraph {
             }
 
             {
-                if !m.end_time.0.is_finite() {
+                if !f64::from(m.end_time).is_finite() {
                     return Err(DemesError::MigrationError(format!(
                         "invalid migration end_time: {:?}",
                         m.end_time
@@ -2921,7 +2658,7 @@ impl UnresolvedGraph {
         self.validate_migrations()?;
 
         match self.generation_time {
-            Some(value) => value.validate()?,
+            Some(value) => value.validate(DemesError::GraphError)?,
             None => {
                 if matches!(self.time_units, TimeUnits::Generations) {
                     self.generation_time = Some(GenerationTime::from(1.));
@@ -2944,7 +2681,7 @@ impl UnresolvedGraph {
 
         if matches!(&self.time_units, TimeUnits::Generations) {
             if let Some(value) = self.generation_time {
-                if !value.0.eq(&1.0) {
+                if !value.eq(&1.0) {
                     return Err(DemesError::GraphError(
                         "time units are generations but generation_time != 1.0".to_string(),
                     ));
@@ -3316,7 +3053,7 @@ mod test_newtype_ordering {
     #[should_panic]
     fn test_fraud_with_start_time() {
         let s = Time::from(1e-3);
-        let sd = Time(f64::NAN);
+        let sd = Time::from(f64::NAN);
         let _ = s < sd;
     }
 }
@@ -3387,7 +3124,7 @@ demes:
         assert_eq!(g.migrations().len(), 1);
         assert_eq!(g.migrations()[0].source(), "A");
         assert_eq!(g.migrations()[0].dest(), "B");
-        assert_eq!(f64::from(g.migrations()[0].rate()), 0.25);
+        assert_eq!(g.migrations()[0].rate(), 0.25);
     }
 
     #[test]
@@ -3413,7 +3150,7 @@ migrations:
         assert_eq!(g.migrations().len(), 1);
         assert_eq!(g.migrations()[0].source(), "A");
         assert_eq!(g.migrations()[0].dest(), "B");
-        assert_eq!(f64::from(g.migrations()[0].rate()), 0.25);
+        assert_eq!(g.migrations()[0].rate(), 0.25);
     }
 
     #[test]
@@ -3439,7 +3176,7 @@ migrations:
         assert_eq!(g.migrations().len(), 1);
         assert_eq!(g.migrations()[0].source(), "A");
         assert_eq!(g.migrations()[0].dest(), "B");
-        assert_eq!(f64::from(g.migrations()[0].rate()), 0.25);
+        assert_eq!(g.migrations()[0].rate(), 0.25);
     }
 
     #[test]
@@ -3465,7 +3202,7 @@ migrations:
         assert_eq!(g.migrations().len(), 1);
         assert_eq!(g.migrations()[0].source(), "A");
         assert_eq!(g.migrations()[0].dest(), "B");
-        assert_eq!(f64::from(g.migrations()[0].rate()), 0.25);
+        assert_eq!(g.migrations()[0].rate(), 0.25);
     }
 
     #[test]
@@ -3491,8 +3228,8 @@ demes:
         assert_eq!(g.migrations()[0].dest(), "B");
         assert_eq!(g.migrations()[1].source(), "B");
         assert_eq!(g.migrations()[1].dest(), "A");
-        assert_eq!(f64::from(g.migrations()[0].rate()), 0.25);
-        assert_eq!(f64::from(g.migrations()[1].rate()), 0.25);
+        assert_eq!(g.migrations()[0].rate(), 0.25);
+        assert_eq!(g.migrations()[1].rate(), 0.25);
     }
 
     #[test]
@@ -3519,8 +3256,8 @@ demes:
         assert_eq!(g.migrations()[0].dest(), "A");
         assert_eq!(g.migrations()[1].source(), "A");
         assert_eq!(g.migrations()[1].dest(), "B");
-        assert_eq!(f64::from(g.migrations()[0].rate()), 0.25);
-        assert_eq!(f64::from(g.migrations()[1].rate()), 0.25);
+        assert_eq!(g.migrations()[0].rate(), 0.25);
+        assert_eq!(g.migrations()[1].rate(), 0.25);
     }
 
     #[test]
@@ -3543,30 +3280,7 @@ demes:
         assert_eq!(g.pulses()[0].sources(), vec!["A".to_string()]);
         assert_eq!(g.pulses()[0].dest(), "B");
         assert_eq!(g.pulses()[0].proportions(), vec![Proportion::from(0.25)]);
-        assert_eq!(f64::from(g.pulses()[0].time()), 100.0);
-    }
-}
-
-#[cfg(test)]
-mod test_infinity {
-    use super::*;
-
-    #[test]
-    fn test_infinity_dot_inf() {
-        let yaml = "---\n.inf\n";
-        let time: Time = serde_yaml::from_str(yaml).unwrap();
-        assert!(time.0.is_infinite());
-        assert!(time.0.is_sign_positive());
-        let yaml = serde_yaml::to_string(&time).unwrap();
-        assert!(yaml.contains("Infinity"));
-    }
-
-    #[test]
-    fn test_infinity_string() {
-        let yaml = "---\nInfinity\n";
-        let time: Time = serde_yaml::from_str(yaml).unwrap();
-        assert!(time.0.is_infinite());
-        assert!(time.0.is_sign_positive());
+        assert_eq!(g.pulses()[0].time(), 100.0);
     }
 }
